requirements.txt
django
djangorestframework
django-filter
Pillow
django-rest-swagger
drf-yasg
